More styling updates
