+++ src/components/LogoutButton.jsx
@@ -1,6 +1,6 @@
 import { useAuth0 } from '@auth0/auth0-react';
 import { Button } from '@mui/material';
-import { ButtonStyle } from './styles';
+import { ButtonStyle } from '../styles';
 
 const LogoutButton = () => {
   const { logout } = useAuth0();
