--- src/components/styles.js
@@ -1,10 +1,0 @@
-export const ButtonStyle = {
-  color: 'white',
-  border: 'none',
-  borderRadius: '30px',
-  backgroundColor: '#EB6159',
-  height: '60px',
-  width: '265px',
-  margin: '0 auto',
-  fontSize: '1rem',
-};
