+++ src/routes/Diagnosis.jsx
@@ -1,9 +1,10 @@
 import { BaseLayout } from '../components/BaseLayout';
+import { SubTitleStyle } from '../styles';
 
 export default function Diagnosis() {
   return (
     <BaseLayout>
-      <h1>Diagnosis</h1>
+      <h1 style={{ alignSelf: 'center', ...SubTitleStyle }}>Diagnosis</h1>
     </BaseLayout>
   );
 }
