+++ src/routes/Information.jsx
@@ -1,9 +1,10 @@
 import { BaseLayout } from '../components/BaseLayout';
+import { SubTitleStyle } from '../styles';
 
 export default function Information() {
   return (
     <BaseLayout>
-      <h1>Information</h1>
+      <h1 style={{ alignSelf: 'center', ...SubTitleStyle }}>Information</h1>
     </BaseLayout>
   );
 }
