+++ src/routes/Share.jsx
@@ -1,9 +1,10 @@
 import { BaseLayout } from '../components/BaseLayout';
+import { SubTitleStyle } from '../styles';
 
 export default function Share() {
   return (
     <BaseLayout>
-      <h1>Share</h1>
+      <h1 style={{ alignSelf: 'center', ...SubTitleStyle }}>Share</h1>
     </BaseLayout>
   );
 }
